homogénéisation size_t

+++ libft/ft_bzero.c
@@ -6,7 +6,7 @@
 /*   By: nedebies <nedebies@student.s19.be>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/03/17 20:33:24 by nedebies          #+#    #+#             */
-/*   Updated: 2021/04/22 14:53:16 by nedebies         ###   ########.fr       */
+/*   Updated: 2021/04/28 12:29:53 by nedebies         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_bzero(void *str, size_t n)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < n)
