+++ libft/ft_calloc.c
@@ -6,7 +6,7 @@
 /*   By: nedebies <nedebies@student.s19.be>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/03/17 20:33:35 by nedebies          #+#    #+#             */
-/*   Updated: 2021/03/17 20:33:40 by nedebies         ###   ########.fr       */
+/*   Updated: 2021/04/28 12:30:12 by nedebies         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 void	*ft_calloc(size_t n, size_t size)
 {
-	char				*ptr;
-	unsigned int		i;
+	char	*ptr;
+	size_t	i;
 
 	i = 0;
 	if (n == 0 || size == 0)
