+++ libft/ft_memccpy.c
@@ -6,7 +6,7 @@
 /*   By: nedebies <nedebies@student.s19.be>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/03/17 20:36:15 by nedebies          #+#    #+#             */
-/*   Updated: 2021/04/05 22:19:14 by nedebies         ###   ########.fr       */
+/*   Updated: 2021/04/28 12:31:07 by nedebies         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memccpy(void *dest, const void *src, int c, size_t n)
 {
-	unsigned int	i;
+	size_t			i;
 	unsigned char	*d;
 	unsigned char	*s;
