+++ libft/ft_memcmp.c
@@ -6,7 +6,7 @@
 /*   By: nedebies <nedebies@student.s19.be>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/03/17 20:36:30 by nedebies          #+#    #+#             */
-/*   Updated: 2021/04/05 22:21:16 by nedebies         ###   ########.fr       */
+/*   Updated: 2021/04/28 12:31:43 by nedebies         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,9 @@
 
 int	ft_memcmp(const void *s1, const void *s2, size_t n)
 {
-	unsigned char			*str1;
-	unsigned char			*str2;
-	unsigned int			i;
+	unsigned char	*str1;
+	unsigned char	*str2;
+	size_t			i;
 
 	i = 0;
 	str1 = (unsigned char *)s1;
