+++ libft/ft_memset.c
@@ -6,7 +6,7 @@
 /*   By: nedebies <nedebies@student.s19.be>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/03/17 20:36:49 by nedebies          #+#    #+#             */
-/*   Updated: 2021/04/05 22:18:24 by nedebies         ###   ########.fr       */
+/*   Updated: 2021/04/28 12:32:40 by nedebies         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 void	*ft_memset(void *str, int c, size_t n)
 {
-	unsigned int	i;
-	char			*arr;
+	size_t	i;
+	char	*arr;
 
 	arr = (char *)str;
 	i = 0;
